Extend functionality for detecting duplicate values in pydantic objects (#105)
* Initial implementation of has_no_duplicate_elements

* renamed find_values_by_name_generic to convey not just pydantic objects

* Simplify parsing tests and use new read/write oscal methods

* Manually edit duplicate uuid's

* add tests for duplicate uuids

* added tools and tests to find duplicate values of specified pydantic type

* better naming of test files and edited uuids

* better implementation of get_values_by_name using fields_set

* Prepare for cli validation

* initial connection to cli

* initial integration of validate test into cli

* Added validation test to cli and refactored
