+++ tests/trestle/cli_test.py
